Added query object to application
