+++ lib/index.js
@@ -2,4 +2,4 @@
 
 exports.Gt = require('./gt'); 
 exports.utils = require('./utils'); 
-// exports.Query = require('./user/query'); 
+exports.Query = require('./query');
