+++ lib/user/create.js
@@ -2,6 +2,7 @@
 
 // const { Validate } = require('../utils');
 const Utils = require('../utils');
+const Query = require('../query');
 const Penseur = require('penseur');
 const Boom = require('boom');
 const Config = require('../config');
@@ -11,63 +12,27 @@ const internals = {};
 
 module.exports.Create = function (userRecord, callback) {
 
-    const createInternals = {
-        action: { name: 'actionName', value: 'actionPayload' }, 
-        db: {}
-    };
-
-    createInternals.action.name = 'validateUserRecord';
-    createInternals.action.value = userRecord;
-
-    const actions = [];
-
-    const validate = function (next, i) {
-
-        Utils.validate.user(userRecord, (err, validatedRecord) => {
-        
-            if (err) {
-                return next(err);
-            }
-
-            createInternals.action.name = 'hashPassword';
-            createInternals.action.value = validatedRecord;
-            return next();
-        });
-    };
-
-
-    actions.push(validate);
-
-    const hashPassword = function (next, i) {
+    Utils.validate.user(userRecord, (err, validatedRecord) => {
     
-        const plainTextPassword = createInternals.action.value.password;
+        if (err) {
+            return callback(err, null);
+        }
+
+        const plainTextPassword = validatedRecord.password;
 
         Utils.bcrypt.hash(plainTextPassword, (err, hashedPassword) => {
 
             if (err) {
-                return next(err);
+                return callback(err, null);
             }
 
-            createInternals.action.name = 'insertRecord';
-            createInternals.action.value.password = hashedPassword;
-            return next();
+            validatedRecord.password = hashedPassword;
+
+            Query.user.create('record', (err, result) => {
+
+                console.log('done create');
+                return callback(null, 'completed items');
+            });
         });
-    };
-
-    actions.push(hashPassword);
-
-
-    const each = (item, next, i) => item(next, i);
-
-    Items.serial(actions, each, (err) => {
-
-        if (err) {
-            // console.log('done Items had error: ' + JSON.stringify(err, null, '\t'));
-            return callback(err, null);
-        }
-
-        // console.log('done Items ' + JSON.stringify(createInternals, null, '\t'));
-        return callback(null, 'completed items');
     });
-
 };
