+++ lib/user/query/create.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports.Create = function (userRecord, callback) {
+
+    console.log('Query.user.create running');
+    return callback(null, 'done');
+};
