+++ lib/utils.js
@@ -7,12 +7,10 @@ const { Compare } = require('./user/utils/bcrypt');
 
 const validate = {};
 validate.user = Validate;
-
 exports.validate = validate; 
 
 
 const bcrypt = {};
 bcrypt.hash = Hash;
 bcrypt.compare = Compare;
-
 exports.bcrypt = bcrypt;
